+++ test/index.integration.test.js
@@ -3,7 +3,6 @@ const rimraf = require('rimraf');
 
 import { copyRecursiveSync, readTextFileSync, readJsonSync, execScriptSync } from './utils/file';
 import { compile, compileTemplate } from './utils/webpack';
-import { regexpAlias, resolveAlias, resolveRequirePath, getResourceParams } from '../src';
 
 // The base path of test directory.
 const basePath = path.resolve(__dirname, './');
@@ -34,11 +33,6 @@ beforeEach(() => {
 });
 
 describe('self tests', () => {
-  it('test it self', (done) => {
-    expect(1).toEqual(1);
-    done();
-  });
-
   it('hello-world', (done) => {
     const relTestCasePath = 'hello-world',
       absTestPath = path.join(PATHS.testOutput, relTestCasePath);
@@ -49,199 +43,6 @@ describe('self tests', () => {
       expect(received).toEqual(expected);
       done();
     });
-  });
-});
-
-describe('resolve alias', () => {
-  const regexp = regexpAlias;
-  const aliases = {
-    App: '/path/to/app/',
-    Component: '/path/to/component/',
-    AppComponent: '/path/to/app-component/',
-  };
-
-  it('App/index.pug', () => {
-    const value = `App/index.pug`;
-    const expected = '/path/to/app/index.pug';
-    const received = resolveAlias(value, aliases, regexp);
-    expect(received).toEqual(expected);
-  });
-
-  it('~App/index.pug', () => {
-    const value = `~App/index.pug`;
-    const expected = '/path/to/app/index.pug';
-    const received = resolveAlias(value, aliases, regexp);
-    expect(received).toEqual(expected);
-  });
-
-  it('@App/index.pug', () => {
-    const value = `@App/index.pug`;
-    const expected = '/path/to/app/index.pug';
-    const received = resolveAlias(value, aliases, regexp);
-    expect(received).toEqual(expected);
-  });
-
-  it('+App/index.pug', () => {
-    const value = `+App/index.pug`;
-    const expected = '+App/index.pug';
-    const received = resolveAlias(value, aliases, regexp);
-    expect(received).toEqual(expected);
-  });
-
-  it('Component/index.pug', () => {
-    const value = `Component/index.pug`;
-    const expected = '/path/to/component/index.pug';
-    const received = resolveAlias(value, aliases, regexp);
-    expect(received).toEqual(expected);
-  });
-
-  it('App/Component/index.pug', () => {
-    const value = `App/Component/index.pug`;
-    const expected = '/path/to/app/Component/index.pug';
-    const received = resolveAlias(value, aliases, regexp);
-    expect(received).toEqual(expected);
-  });
-
-  it('App-Component/index.pug', () => {
-    const value = `App-Component/index.pug`;
-    const expected = 'App-Component/index.pug';
-    const received = resolveAlias(value, aliases, regexp);
-    expect(received).toEqual(expected);
-  });
-
-  it('App_Component/index.pug', () => {
-    const value = `App_Component/index.pug`;
-    const expected = 'App_Component/index.pug';
-    const received = resolveAlias(value, aliases, regexp);
-    expect(received).toEqual(expected);
-  });
-
-  it('/path/to/App/index.pug', () => {
-    const value = `/path/to/App/index.pug`;
-    const expected = '/path/to/App/index.pug';
-    const received = resolveAlias(value, aliases, regexp);
-    expect(received).toEqual(expected);
-  });
-});
-
-describe('resolve embedded resource', () => {
-  const templateFile = '/path/to/source/template.pug';
-  const resolveAliases = {
-    App: '/path/to/app/',
-    Component: '/path/to/component/',
-    AppComponent: '/path/to/app-component/',
-  };
-
-  it(`img(src=require('./image.jpeg'))`, () => {
-    const value = `require('./image.jpeg')`;
-    const expected = `require('/path/to/source/' + 'image.jpeg')`;
-    const received = resolveRequirePath(value, templateFile, resolveAliases);
-    expect(received).toEqual(expected);
-  });
-
-  it(`img(src=require("./image.jpeg"))`, () => {
-    const value = `require("./image.jpeg")`;
-    const expected = `require('/path/to/source/' + "image.jpeg")`;
-    const received = resolveRequirePath(value, templateFile, resolveAliases);
-    expect(received).toEqual(expected);
-  });
-
-  it(`img(src=require('../image.jpeg'))`, () => {
-    const value = `require('../image.jpeg')`;
-    const expected = `require('/path/to/source/' + '../image.jpeg')`;
-    const received = resolveRequirePath(value, templateFile, resolveAliases);
-    expect(received).toEqual(expected);
-  });
-
-  it(`img(src=require('App/image.jpeg'))`, () => {
-    const value = `require('App/image.jpeg')`;
-    const expected = `require('/path/to/app/image.jpeg')`;
-    const received = resolveRequirePath(value, templateFile, resolveAliases);
-    expect(received).toEqual(expected);
-  });
-
-  it(`img(src=require('Component/image.jpeg'))`, () => {
-    const value = `require('Component/image.jpeg')`;
-    const expected = `require('/path/to/component/image.jpeg')`;
-    const received = resolveRequirePath(value, templateFile, resolveAliases);
-    expect(received).toEqual(expected);
-  });
-
-  it(`img(src=require('App/Component/image.jpeg'))`, () => {
-    const value = `require('App/Component/image.jpeg')`;
-    const expected = `require('/path/to/app/Component/image.jpeg')`;
-    const received = resolveRequirePath(value, templateFile, resolveAliases);
-    expect(received).toEqual(expected);
-  });
-
-  it(`img(src=require('AppTest/image.jpeg'))`, () => {
-    const value = `require('AppTest/image.jpeg')`;
-    const expected = `require('/path/to/source/' + 'AppTest/image.jpeg')`;
-    const received = resolveRequirePath(value, templateFile, resolveAliases);
-    expect(received).toEqual(expected);
-  });
-
-  it(`img(src=require('App-Component/image.jpeg'))`, () => {
-    const value = `require('App-Component/image.jpeg')`;
-    const expected = `require('/path/to/source/' + 'App-Component/image.jpeg')`;
-    const received = resolveRequirePath(value, templateFile, resolveAliases);
-    expect(received).toEqual(expected);
-  });
-
-  it(`img(src=require('App_Component/image.jpeg'))`, () => {
-    const value = `require('App_Component/image.jpeg')`;
-    const expected = `require('/path/to/source/' + 'App_Component/image.jpeg')`;
-    const received = resolveRequirePath(value, templateFile, resolveAliases);
-    expect(received).toEqual(expected);
-  });
-
-  it(`require(path + '')`, () => {
-    const value = `require(path + '')`;
-    const expected = `require('/path/to/source/' + path + '')`;
-    const received = resolveRequirePath(value, templateFile, resolveAliases);
-    expect(received).toEqual(expected);
-  });
-
-  it(`require('' + path)`, () => {
-    const value = `require('' + path)`;
-    const expected = `require('/path/to/source/' + '' + path)`;
-    const received = resolveRequirePath(value, templateFile, resolveAliases);
-    expect(received).toEqual(expected);
-  });
-
-  it(`require('' + path + '')`, () => {
-    const value = `require('' + path + '')`;
-    const expected = `require('/path/to/source/' + '' + path + '')`;
-    const received = resolveRequirePath(value, templateFile, resolveAliases);
-    expect(received).toEqual(expected);
-  });
-
-  it('img(src=require(`${path}`))', () => {
-    const value = 'require(`${path}`)';
-    const expected = "require('/path/to/source/' + `${path}`)";
-    const received = resolveRequirePath(value, templateFile, resolveAliases);
-    expect(received).toEqual(expected);
-  });
-
-  it('require(`App/images/${file}`)', () => {
-    const value = 'require(`App/images/${file}`)';
-    const expected = 'require(`/path/to/app/images/${file}`)';
-    const received = resolveRequirePath(value, templateFile, resolveAliases);
-    expect(received).toEqual(expected);
-  });
-
-  it('require(`../../../images/${file}`)', () => {
-    const value = 'require(`../../../images/${file}`)';
-    const expected = "require('/path/to/source/' + '../../../images/' + `${file}`)";
-    const received = resolveRequirePath(value, templateFile, resolveAliases);
-    expect(received).toEqual(expected);
-  });
-
-  it('img(src=require(`App/${path}`))', () => {
-    const value = 'require(`App/${path}`)';
-    const expected = 'require(`/path/to/app/${path}`)';
-    const received = resolveRequirePath(value, templateFile, resolveAliases);
-    expect(received).toEqual(expected);
   });
 });
 
@@ -503,8 +304,8 @@ describe('require pug in javascript', () => {
     });
   });
 
-  it(`render embedded resources`, (done) => {
-    const relTestCasePath = 'javascript-render-embedded-resources',
+  it(`javascript compile embedded resources`, (done) => {
+    const relTestCasePath = 'javascript-assets-compile',
       absTestPath = path.join(PATHS.testOutput, relTestCasePath);
 
     compile(PATHS, relTestCasePath).then(() => {
@@ -515,120 +316,17 @@ describe('require pug in javascript', () => {
       done();
     });
   });
-});
 
-describe('parse resource data', () => {
-  it('empty string', () => {
-    const expected = {};
-    const received = getResourceParams('');
-    expect(received).toEqual(expected);
-  });
+  it(`javascript render embedded resources`, (done) => {
+    const relTestCasePath = 'javascript-assets-render',
+      absTestPath = path.join(PATHS.testOutput, relTestCasePath);
 
-  it('?', () => {
-    const expected = {};
-    const received = getResourceParams('?');
-    expect(received).toEqual(expected);
-  });
+    compile(PATHS, relTestCasePath).then(() => {
+      const received = execScriptSync(path.join(absTestPath, PATHS.assets, 'index.js'));
+      const expected = readTextFileSync(path.join(absTestPath, PATHS.expected, 'output.html'));
 
-  it('?pug_method', () => {
-    const expected = {
-      pug_method: '',
-    };
-    const received = getResourceParams('?pug_method');
-    expect(received).toEqual(expected);
-  });
-
-  it('?pug_method=', () => {
-    const expected = {
-      pug_method: '',
-    };
-    const received = getResourceParams('?pug_method=');
-    expect(received).toEqual(expected);
-  });
-
-  it('?pug_method=&', () => {
-    const expected = {
-      pug_method: '',
-    };
-    const received = getResourceParams('?pug_method=&');
-    expect(received).toEqual(expected);
-  });
-
-  it('?pug_method=render', () => {
-    const expected = {
-      pug_method: 'render',
-    };
-    const received = getResourceParams('?pug_method=render');
-    expect(received).toEqual(expected);
-  });
-
-  it('?pug_method=render&', () => {
-    const expected = {
-      pug_method: 'render',
-    };
-    const received = getResourceParams('?pug_method=render&');
-    expect(received).toEqual(expected);
-  });
-
-  /*it('?pug_render=true', () => {
-    const expected = {
-      pug_render: true,
-    };
-    const received = getResourceParams('?pug_render=true');
-    expect(received).toEqual(expected);
-  });*/
-
-  /*it('?a=123', () => {
-    const expected = {
-      a: 123,
-    };
-    const received = getResourceParams('?a=123');
-    expect(received).toEqual(expected);
-  });*/
-
-  it('?{"a":10,"b":"abc"}', () => {
-    const expected = {
-      a: 10,
-      b: 'abc',
-    };
-    const received = getResourceParams('?{"a":10,"b":"abc"}');
-    expect(received).toEqual(expected);
-  });
-
-  it('?&{"a":10,"b":"abc"}', () => {
-    const expected = {
-      a: 10,
-      b: 'abc',
-    };
-    const received = getResourceParams('?&{"a":10,"b":"abc"}');
-    expect(received).toEqual(expected);
-  });
-
-  it('?pug_method=render&{"a":10,"b":"abc"}', () => {
-    const expected = {
-      pug_method: 'render',
-      a: 10,
-      b: 'abc',
-    };
-    const received = getResourceParams('?pug_method=render&{"a":10,"b":"abc"}');
-    expect(received).toEqual(expected);
-  });
-
-  it('?pug_method=render&opts={"a":10,"b":"abc"}', () => {
-    const expected = {
-      pug_method: 'render',
-      options: { a: 10, b: 'abc' },
-    };
-    const received = getResourceParams('?pug_method=render&options={"a":10,"b":"abc"}');
-    expect(received).toEqual(expected);
-  });
-
-  it('?pug_method=render&opts[]=a1&opts[]=a2', () => {
-    const expected = {
-      pug_method: 'render',
-      args: ['a1', 'a2'],
-    };
-    const received = getResourceParams('?pug_method=render&args[]=a1&args[]=a2');
-    expect(received).toEqual(expected);
+      expect(received).toEqual(expected);
+      done();
+    });
   });
 });