+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change log
 
+## 1.3.0 (2021-11-15)
+- feature: the `render` method has been improved. Now the method render a pug into HTML really at compile time without limitations for resolving an embedded resource.
+  This method do same result as any other pug-loader + html-loader, even faster, generate smaller code and with all that not need an additional loader.
+- some code and test refactorings
+
 ## 1.2.0 (2021-11-12)
 - feature: added for the loader option `method` new value `html` to render the template function into pure HTML string,\
   this method require additional loader, e.g. `html-loader`, see usage sample in readme